backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
sentence-transformers==2.3.1
rank-bm25==0.2.2
spacy==3.7.2
numpy==1.26.3
faiss-cpu==1.7.4
networkx==3.2.1
openai==1.10.0
requests==2.31.0
supabase==2.9.0
